s2mpg1x-regulator: PMIC read/write support
Bug: 170918051

+++ include/linux/mfd/samsung/s2mpg10.h
@@ -228,7 +228,7 @@ struct s2mpg10_pmic {
 #if IS_ENABLED(CONFIG_DRV_SAMSUNG_PMIC)
 	struct device *dev;
 	u8 read_val;
-	u8 read_addr;
+	u16 read_addr;
 #endif
 };
